Consulta de menu y primera parte de generacion de ingredientes de menu

+++ mf-proy/mf-service-menu/src/main/java/pe/com/rhsistemas/mfsermenu/service/impl/MenuLogicaServiceImpl.java
@@ -379,7 +379,7 @@ public class MenuLogicaServiceImpl implements MenuLogicaService {
 		Integer idPlatoElegido = listaPlatosNoConsumidos.get((numeroElegido == 0 ? 1 : numeroElegido) - 1).getId();
 		
 		menuDia.getPlatoDto().setId(idPlatoElegido);
-		menuDia.setIdGenerado(Long.valueOf(ultimoMenuGenerado.getIdGenerado()).intValue());
+		menuDia.setIdGenerado(ultimoMenuGenerado.getIdGenerado());
 		
 		log.info("fecha consumo :: "+menuDia.getFechaConsumo());
